pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "researchswarm-registry"
version = "0.1.0"
description = "Tool Registry Service for the Research Swarm system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "sentence-transformers>=3.0",
    "google-genai>=1.0",
    "httpx>=0.28",
    "structlog>=24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "alembic>=1.14",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "black>=24.0",
    "coverage>=7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.setuptools.packages.find]
include = ["registry*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
target-version = ["py311"]
line-length = 100
